server/requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
PyPDF2
Pillow
sse-starlette
stripe
supabase
